requirements.txt
alembic==1.18.4
annotated-types==0.7.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.3.2
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.30.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
pandas==3.0.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdf==6.7.2
python-dateutil==2.8.2
python-dotenv==1.1.1
reportlab==4.4.3
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.47
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
